setup frontend routing, login and session
