+++ .env.example
@@ -1,1 +1,2 @@
 ENV=dev
+APP_KEY=
