+++ package.json
@@ -10,8 +10,9 @@
   },
   "dependencies": {
     "axios": "^1.6.2",
-    "bootstrap": "^5.3.2",
-    "preact": "^10.19.1"
+    "bootstrap": "^4.6.2",
+    "preact": "^10.19.1",
+    "preact-router": "^4.1.2"
   },
   "devDependencies": {
     "@preact/preset-vite": "^2.6.0",
