+++ run.sh
@@ -1,0 +1,1 @@
+flask --app router run --debug
