--- templates/home.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <link rel="stylesheet" href="/node_modules/bootstrap/dist/css/bootstrap.min.css">
-  <title>Home</title>
-</head>
-<body>
-  <div class="row">
-    <div class="col">
-      <input>
-    </div>
-    <div class="col">
-      <input>
-    </div>
-  </div>
-</body>
-</html>
