+++ vite.config.js
@@ -4,4 +4,13 @@ import preact from '@preact/preset-vite'
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [preact()],
+  server: {
+    proxy: {
+      '/api': {
+        target: 'http://127.0.0.1:5000/',
+        changeOrigin: true,
+        // rewrite: (path) => path.replace(/^\/api/, ''),
+      }
+    },
+  },
 })
